requirements.txt
@"
# Core
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23

# Desktop
pywebview==4.4.1

# API Clients
httpx==0.25.1
aiohttp==3.9.0
requests==2.31.0

# Image Processing
Pillow==10.1.0

# Security (without bcrypt)
cryptography==41.0.7
passlib==1.7.4

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
PyYAML==6.0.1
aiofiles==23.2.1
"@ | Out-File -FilePath requirements_windows.txt -Encoding UTF